+++ front/src/app/modules/admin-course/interfaces/course-display.ts
@@ -2,9 +2,15 @@ export interface IPortDisplay {
   type: string;
 }
 
+export enum CustomDisplayType {
+  BUTTON = 'BOUTON',
+  TEXT = 'TEXTE'
+}
+
 export interface IHttpPortDisplay extends IPortDisplay {
   url: string;
   title: string;
+  customDisplayType: CustomDisplayType
 }
 
 export interface ICourseDisplay {