+++ front/src/app/modules/display-container/render-string.service.spec.ts
@@ -33,7 +33,7 @@ describe("RenderStringService", () => {
     description: "http",
     mapPort: 8889,
     portMapTo: 8889,
-    connectionType: ConnectionType.OTHER,
+    connectionType: ConnectionType.TCP,
     requiredToAccessContainer: false,
   }];